More changes

+++ controllers/cards.ts
@@ -1,13 +1,12 @@
-import { UserRequest } from "../interface/users";
 import Card from "../models/card";
 import { NextFunction, Request, Response } from "express";
 import { NotFoundError } from "../errors/not-found-error";
 
-export const createCard = (req: UserRequest, res: Response, next: NextFunction) => {
+export const createCard = (req: Request, res: Response, next: NextFunction) => {
   return Card.create({
     name: req.body.name,
     link: req.body.link,
-    owner: req.user!._id,
+    owner: req.user,
   })
     .then((card) => res.status(201).send(card))
     .catch(next);
@@ -20,6 +19,10 @@ export const getCards = (req: Request, res: Response, next: NextFunction) => {
 };
 
 export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
+  if (req.user !== req.params.owner) {
+    return res.status(403).send({ message: "Нельзя удалить чужую карту" });
+  }
+
   return Card.findByIdAndDelete(req.params.cardId)
     .then((card) => {
       if (!card) {
@@ -30,10 +33,10 @@ export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
     .catch(next);
 };
 
-export const addLike = (req: UserRequest, res: Response, next: NextFunction) => {
+export const addLike = (req: Request, res: Response, next: NextFunction) => {
   return Card.findByIdAndUpdate(
     req.params.cardId,
-    { $addToSet: { likes: req.user!._id } },
+    { $addToSet: { likes: req.user } },
     { new: true }
   )
     .then((card) => {
@@ -45,10 +48,10 @@ export const addLike = (req: UserRequest, res: Response, next: NextFunction) => 
     .catch(next);
 };
 
-export const removeLike = (req: UserRequest, res: Response, next: NextFunction) => {
+export const removeLike = (req: Request, res: Response, next: NextFunction) => {
   return Card.findByIdAndUpdate(
     req.params.cardId,
-    { $pull: { likes: req.user!._id } },
+    { $pull: { likes: req.user } },
     { new: true }
   )
     .then((card) => {
